Big push... integrated conversion code into main module for simplification, and will probably move it back out of main once I start writing tests. Implemented a tkinter askopenfilename() gui to select the file path, and cleaned up some var names.
Created a temp directory in the project root to store a tempCSV.csv file. There's probably a better way around having to write to a temp file but I haven't figured that out yet - I tried using an SO hint that converts a list to a pandas dataframe and then into a csv file, but that mucked up the final csv big time so I just returned to my old clunky way of doing it.

Testing will happen tomorrow (11/20) and then I'll start cleaning at the end of the day or starting the following day.